Restructure photos files in preparation of upcoming updates
