+++ src/components/BatchWithdrawFlow/flow/AddPhotos.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { Container, useTheme } from '@mui/material';
 
 import PageForm from 'src/components/common/PageForm';
-import SelectPhotos from 'src/components/common/SelectPhotos';
+import SelectPhotos from 'src/components/common/Photos/SelectPhotos';
 import strings from 'src/strings';
 import useDeviceInfo from 'src/utils/useDeviceInfo';
